requirements.txt
streamlit
opencv-python
face_recognition
numpy
pandas
pymysql
dlib
pytz
